+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* ability to include more edge/node tags into the graph
+### Changed
+
+* dropped unused oneway convertion
+
 ## [0.1.0] - 2020-01-18
 
 ### Added